Conversion des objets Xee vers des objets sauvegardables en base de données via RoomDatabase. Les services Xee ne marchent plus sans aucune raison apparente, check le code dans SettingsFragment

+++ app/src/main/java/com/nsy209/nicedriver/model/AppDatabase.java
@@ -5,18 +5,15 @@ import android.arch.persistence.room.Room;
 import android.arch.persistence.room.RoomDatabase;
 import android.content.Context;
 import android.os.Environment;
-import android.util.Log;
 
-import com.nsy209.nicedriver.model.dao.MarkerDao;
-import com.nsy209.nicedriver.model.dao.PointDao;
-import com.nsy209.nicedriver.model.dao.ReferentielDao;
-import com.nsy209.nicedriver.model.dao.UserDao;
+import com.nsy209.nicedriver.model.dao.LocationDao;
+import com.nsy209.nicedriver.model.dao.SignalDao;
+import com.nsy209.nicedriver.model.dao.TripDao;
 import com.nsy209.nicedriver.model.objects.Car;
+import com.nsy209.nicedriver.model.objects.Location;
 import com.nsy209.nicedriver.model.objects.Marker;
-import com.nsy209.nicedriver.model.objects.Point;
-import com.nsy209.nicedriver.model.objects.Referentiel;
+import com.nsy209.nicedriver.model.objects.Signal;
 import com.nsy209.nicedriver.model.objects.Trip;
-import com.nsy209.nicedriver.model.objects.User;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -27,20 +24,16 @@ import java.nio.channels.FileChannel;
  * Created by Sébastien on 21/07/2017.
  */
 
-@Database(entities = {Car.class, Marker.class, Point.class, Referentiel.class, Trip.class, User.class}, version = 1)
+@Database(entities = {Car.class, Marker.class, Trip.class, Location.class, Signal.class}, version = 1)
 @android.arch.persistence.room.TypeConverters({TypeConverters.class})
 public abstract class AppDatabase extends RoomDatabase {
     private static final String DATABASE_NAME = "nicedriver-database";
     private static final String TAG = AppDatabase.class.getName();
     private static AppDatabase INSTANCE;
 
-    public abstract MarkerDao markerDao();
-
-    public abstract PointDao pointDao();
-
-    public abstract ReferentielDao referentielDao();
-
-    public abstract UserDao userDao();
+    public abstract TripDao tripDao();
+    public abstract SignalDao signalDao();
+    public abstract LocationDao locationDao();
 
     public static AppDatabase getAppDatabase(Context context) {
         if (INSTANCE == null) {
@@ -50,30 +43,30 @@ public abstract class AppDatabase extends RoomDatabase {
         return INSTANCE;
     }
 
-    public static boolean exportDatabase(Context context) {
+
+    public static void exportDatabase(Context context, String name) {
         try {
             File sd = Environment.getExternalStorageDirectory();
-//            File data = context.getDatabasePath(DATABASE_NAME);
 
             if (sd.canWrite()) {
-//                String currentDBPath = "//data//" + "com.nsy209.nicedriver"
-//                        + "//databases//" + DATABASE_NAME;
-                String backupDBPath = DATABASE_NAME + ".db";
-//                File currentDB = new File(data, currentDBPath);
-                File currentDB = context.getDatabasePath(DATABASE_NAME);
+                String currentDBPath = context.getDatabasePath(DATABASE_NAME).getAbsolutePath();
+                String backupDBPath = name;
+                //previous wrong  code
+                // **File currentDB = new File(data,currentDBPath);**
+                // correct code
+                File currentDB = new File(currentDBPath);
                 File backupDB = new File(sd, backupDBPath);
 
-                FileChannel src = new FileInputStream(currentDB).getChannel();
-                FileChannel dst = new FileOutputStream(backupDB).getChannel();
-                dst.transferFrom(src, 0, src.size());
-                src.close();
-                dst.close();
-                Log.d(TAG, "Export database success");
+                if (currentDB.exists()) {
+                    FileChannel src = new FileInputStream(currentDB).getChannel();
+                    FileChannel dst = new FileOutputStream(backupDB).getChannel();
+                    dst.transferFrom(src, 0, src.size());
+                    src.close();
+                    dst.close();
+                }
             }
-            return true;
         } catch (Exception e) {
             e.printStackTrace();
-            return false;
         }
     }
 }
